moved into seperate feature package and added config to disable
