+++ common/src/main/java/org/valkyrienskies/mod/mixin/feature/block_tint/FEATURE.md
@@ -1,0 +1,8 @@
+# Block Tinting
+
+These mixins fix block tinting for blocks on ships.
+
+(Only updates if the chunk render buffer is updated)
+
+Adds a client side option in the config (called "fixBlockTinting" in the category "BlockTinting") to enable this feature.
+(default off!)
